Extract StateBasedInteractor
