+++ frontend/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './HTTPClient';
 export * from './sleep';
 export * from './validations';
 export * from './object';
+export * from './StateBasedInteractor';
